# api/tmp1/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: pharmacies
  description: pharmacies tasks
  contact:
    email: abdoulfataoh@gmail.com
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: test server
paths:
  /health:
    get:
      summary: Get server health status
      operationId: health_get
      responses:
        "200":
          description: server status
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: 1430401802
                result: ok
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /products/list:
    get:
      summary: Returns a list of products
      operationId: products_list_get
      responses:
        "200":
          description: A JSON array of products
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: "1430401802,"
                result:
                - _id: 614cc69cac9eed1f9bde8542
                  nom_commercial: abaisse langue en bois
                  dci: []
                  conditionnement: b/100
                  prix_public: 838
                - _id: 614cc69cac9eed1f9bde854c
                  nom_commercial: bactyl
                  dci:
                  - céthexonium
                  dosage: 2.5E-4
                  forme_galemique: collyre
                  conditionnement: fl/10ml
                  prix_public: 892
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /products/add:
    post:
      summary: add a product
      operationId: products_add_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/products_add_body'
        required: true
      responses:
        "200":
          description: reponse status
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: "1430401802,"
                result: true
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /products/{product_id}:
    delete:
      summary: delete a product
      operationId: products_product_id_delete
      parameters:
      - name: product_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: reponse status
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: "1430401802,"
                result: true
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /pharmacies/list:
    get:
      summary: Returns a list of pharmacies
      operationId: pharmacies_list_get
      responses:
        "200":
          description: A JSON array of pharmacies
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: "1430401802,"
                result:
                - _id: 614d0e0999690cb6123f53f8
                  nom_pharmacie: avenir
                  telephone: 25361338
                  code_postale: 01 bp 4881
                  addresse: secteur 14 (zogona)
                  assurance:
                  - sonar
                  - ' sunu'
                  - ' uab'
                  groupe: 1
                  products_ids: []
                  localisation:
                    type: Point
                    coordinates:
                    - -1.49388
                    - 12.37513
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /pharmacie/add:
    post:
      summary: add a pharmacie
      operationId: pharmacie_add_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pharmacie_add_body'
        required: true
      responses:
        "200":
          description: reponse status
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: "1430401802,"
                result: true
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /pharmacie/{pharmacie_id}:
    delete:
      summary: delete a pharmacie by id
      operationId: pharmacie_pharmacie_id_delete
      parameters:
      - name: pharmacie_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: reponse status
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: "1430401802,"
                result: true
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /pharmacies/proximities/list:
    get:
      summary: Returns a list of pharmacies order by proximity
      operationId: pharmacies_proximities_list_get
      parameters:
      - name: lat
        in: query
        description: client location latitude
        required: true
        style: form
        explode: true
        schema:
          type: number
      - name: long
        in: query
        description: client location longitude
        required: true
        style: form
        explode: true
        schema:
          type: number
      responses:
        "200":
          description: A order by  JSON array of pharmacies
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: "1430401802,"
                result:
                - _id: 614c67ec7831965969347dc6
                  nom_pharmacie: avenir
                  telephone: 25361338
                  code_postale: 01 bp 4881
                  addresse: secteur 14 (zogona)
                  assurance:
                  - sonar
                  - ' sunu'
                  - ' uab'
                  groupe: groupe 1
                  localisation:
                    type: Point
                    coordinates:
                    - -1.49388
                    - 12.37513
                  distance: 9496.179816887958
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /pharmacies/mananger/{pharmacie_id}/set_available_products:
    post:
      summary: set products list as available products for a pharmacie
      operationId: pharmacies_mananger_pharmacie_id_set_available_products_post
      parameters:
      - name: pharmacie_id
        in: path
        description: pharmacie_id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pharmacie_id_set_available_products_body'
        required: true
      responses:
        "200":
          description: reponse status
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: "1430401802,"
                result: true
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /pharmacies/manager/{pharmacie_id}/set_unvailable_products:
    post:
      summary: set products list as available products for a pharmacie
      operationId: pharmacies_manager_pharmacie_id_set_unvailable_products_post
      parameters:
      - name: pharmacie_id
        in: path
        description: pharmacie_id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pharmacie_id_set_unvailable_products_body'
        required: true
      responses:
        "200":
          description: reponse status
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: "1430401802,"
                result: true
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /pharmacies/proximities/{search_product_id}:
    post:
      summary: Returns a list of pharmacies order by proximity with a specic product
      operationId: pharmacies_proximities_search_product_id_post
      parameters:
      - name: search_product_id
        in: path
        description: product id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/search_pharmacies_body'
        required: true
      responses:
        "200":
          description: A order by  JSON array of pharmacies
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
              example:
                success: "true,"
                timestamp: "1430401802,"
                result:
                - _id: 614c67ec7831965969347dc6
                  nom_pharmacie: avenir
                  telephone: 25361338
                  code_postale: 01 bp 4881
                  addresse: secteur 14 (zogona)
                  assurance:
                  - sonar
                  - ' sunu'
                  - ' uab'
                  groupe: groupe 1
                  products_ids:
                  - 614b9bbe257b7ec7226aad2d
                  - 614b9bbe257b7ec7226aad2d
                  localisation:
                    type: Point
                    coordinates:
                    - -1.49388
                    - 12.37513
                  distance: 9496.179816887958
      x-openapi-router-controller: swagger_server.controllers.default_controller
components:
  schemas:
    products_add_body:
      required:
      - conditionnement
      - dci
      - dosage
      - forme_galemique
      - lab_fabricant
      - nom_commercial
      - prix_public
      type: object
      properties:
        nom_commercial:
          type: string
        dci:
          type: array
          items:
            type: string
        dosage:
          type: string
        forme_galemique:
          type: string
        lab_fabricant:
          type: string
        conditionnement:
          type: array
          items:
            type: string
        prix_public:
          type: number
    pharmacie_add_body:
      required:
      - localisation
      - nom_pharmacie
      type: object
      properties:
        nom_pharmacie:
          type: string
        telephone:
          type: string
        code_postale:
          type: string
        addresse:
          type: string
        assurance:
          type: array
          items:
            type: string
        groupe:
          type: number
        localisation:
          type: array
          items:
            type: number
        products_ids:
          type: array
          items:
            type: string
    pharmacie_id_set_available_products_body:
      required:
      - pharmacie_id
      - products_ids
      type: object
      properties:
        pharmacie_id:
          type: string
        products_ids:
          type: array
          items:
            type: string
    pharmacie_id_set_unvailable_products_body:
      required:
      - pharmacie_id
      - products_ids
      type: object
      properties:
        pharmacie_id:
          type: string
        products_ids:
          type: array
          items:
            type: string
    search_pharmacies_body:
      required:
      - lat
      - long
      type: object
      properties:
        lat:
          type: number
        long:
          type: number
        assurance:
          type: array
          items:
            type: string